+++ packages/common/src/api/index.ts
@@ -161,6 +161,7 @@ export * from './tan-query/wallets/useWalletOwner'
 export * from './tan-query/wallets/useUSDCBalance'
 export * from './tan-query/wallets/useExternalWalletBalance'
 export * from './tan-query/wallets/useCoinBalance'
+export * from './tan-query/wallets/useCoinBalanceBreakdown'
 export * from './tan-query/wallets/useSendCoins'
 export * from './tan-query/jupiter/useSwapCoins'
 export * from './tan-query/jupiter/useCoinExchangeRate'
